Code refactoring according to comments
